Several updates
Removed error message to user within data input validation methods; message now must be displayed in the module interacting with the user and employing the framework. This makes much more sense and doesn't restrict the error message to the user to whatever the framework is set up for. Also added a new Exception class to be used to ensure lists/tuples are the same length (e.g., when they will be zipped). Miscellaneous other improvements and comment updates.